copy starter code from feather-tft-gamepad
This is a copy and paste from my feather-tft-gamepad repo with
edits to remove stuff that's not relevant for Fruit Jam hardware.

Haven't tested this code yet. It's probably broken. Need to get the
display code tuned for Fruit Jam. Might need to change how the USB
code works as the CircuitPython USB stack implementation changed a
lot recently.