working on usb descriptor parser
CircuitPython's subset of the PyUSB API does not include a user
accessible parser for device, configuration, interface, or endpoint
descriptors. I need that information to identify DInput and
XInput gamepads without having to build a giant list of vid:pid
pairs in advance. So, I'm making my own parser.

This is the start of what I hope will be an automatic detection
system to identify XInput or DInput gamepads regardless of their
vendor or product ID. Not sure if this will work, but I'm gonna
give it a solid attempt.