overhaul timeout behavior + Button #1 for rescan
This significantly improves initial connection reliability at the
expense of not being able to detect when the USB device has been
unplugged. Now you have to press Button #1 to manually initiate a
USB bus re-scan.

As far as I can tell, there's currently (10.0.0-alpha.4) no way to
distinguish between a USBTimeoutError that happens when the gamepad
is operating normally and a timeout error that happens when the
device has been unplugged.

Previously, I assumed that a run of timeout errors meant the device
had been disconnected so I should re-scan the USB bus. That worked
well to detect unplug events, but it made the initialization
glitchy and the connection unstable for one of my gamepads.

After this change, the initial connect is much faster, and the
connection seems rock solid for both gamepads I've been testing
with.