wip: looking at HID report descriptors
For now, this just hexdumps the HID report descriptor when it finds
a device with generic HID on interface 0. Looks like about the
first 6 bytes of the report descriptor should be sufficient to
fingerprint DInput gamepads vs. Switch-compatible controllers. Not
sure if I should build a full parser. Don't have a frame of
reference for the amount of variation across DInput controllers.