rework descriptor pretty printing
This commit is mostly about reducing log spam. Net effect of the
changes is to make it so that the configuration descriptor for a
USB device only gets printed once rather than many times in a row.

Previously, devices that didn't pass the is-it-a-gamepad filter
would get scanned and printed each time the retry loop initiated
another check for gamepads.