allow longer HID report descriptors
The longest HID report descriptor I've seen among the devices I
have on hand to test with is 400+ bytes for what appears to be a
multi-touch trackpad interface. This change allows for up to 512
bytes read by ctrl_transfer(), but it only allocates the size
indicated by the configuration descriptor.