show more device details on DVI GUI
This expands the USB device metadata display to include vid:pid,
detected device type tag, device class:subclass:protocol, and
interface 0 class:subclass:protocol.

The main effective difference is that it now detects and displays
boot mice and boot keyboards on the GUI, in addition to gamepads.

The data model and class hierarchy I'm using probably need to be
overhauled a bit to deal with the concept of scan results that can
be various types of input devices. I started this as a strictly
gamepad thing, but now I want the code to be aware of keyboards and
mice too.