overhaul USB input event polling
This reworks USB input event polling to use a generator function
that makes an iterator which yields a bitfield of button events.
There's also a new mechanism to throttle USB endpoint reads to
avoid spamming the USB bus with requests that are too frequent for
the device to respond to.

The point of doing it this way is:
1. Reduce heap allocation pressure and dictionary lookups
2. Allow for multiple backend USB device drivers to share frontend
   API without the caller needing to know implementation details
3. Try to improve USB IO stability (avoid CPU lockup & reset)