upgrade to 10.0.0-alpha.5, fix XInput init
Just upgraded my Fruit Jam with CircuitPython 10.0.0-alpha.5 which
came out yesterday.

Initial impression: USB seems better behaved. From a quick test, it
seems that unplugging a device is more likely to give me a USBError
now, where it would often do USBTimeoutError previously. This may
relate to the latest TinyUSB incorporating a workaround for the
RP2350 E9 erratum on the USB data pins? Think maybe I saw a mention
of that in an issue or PR thread somewhere? Not sure.

Need to test more, but I hope this will let me reliably detect
unplugs with an error that looks different from the regular
timeouts. If so, I can drop my press-button-1-to-rescan-usb bodge,
which I would be happy to be rid of.