hexdump HID reports to console + DVI
This shows HID reports from mice, keyboards, and HID gamepads as
a hexdump on the console and DVI display output.

It works, but it's slow. Main problem is probably that each change
to the input report triggers half a screen worth of text rendering.
Lots of room to make that more efficient, but this is a good proof
of concept. Now I kinda want add an option for converting the HID
report data to a bitmap to better show what's changing inside long
gamepad reports.

Looking at how this works makes me think that perhaps I should
forget about button mapping for DInput gamepads. It might be a lot
more straightforward to just give people an API to detect the type
of device and to get raw HID report data. They can deal with the
button mapping on their own. It's pretty easy to pull out the bits
and bytes you need once you have a visualization to show how the
data changes in response to control inputs.